third_party/blink/web_tests/virtual/shared_array_buffer_on_desktop/fast/workers/worker-atomics-wait-expected.txt

Test that Atomics.wait is not allowed on the main thread.

PASS: Calling Atomics.wait on the main thread throws.
PASS: Atomics.wait is allowed in a worker.
DONE
